Refine failure explanation projection
